chore: disable react-router-dom singleton (#14)
consoledot needs this to be off to be able to upgrade the router to v6. We don't need this to be a singleton, so let's keep this off. Also, remove all eager loading as we should never eagerly load any dependency ever; when eagerly loading, the dependency gets bundled with the bundle that will be loaded through module federation, forcing users to download them all over again